requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
starlette==0.37.2
pymongo==4.7.2
python-dotenv==1.0.1
google-generativeai==0.5.4
python-dateutil==2.9.0
marshmallow==3.21.1
